changed from Int to Long as the content length in bytes could be large

+++ zio-http/src/main/scala/zio/http/model/headers/values/ContentLength.scala
@@ -13,7 +13,7 @@ object ContentLength {
    * The Content-Length header indicates the size of the message body, in bytes,
    * sent to the recipient.
    */
-  final case class ContentLengthValue(length: Int) extends ContentLength
+  final case class ContentLengthValue(length: Long) extends ContentLength
 
   /**
    * The ContentLength header value is invalid.
@@ -27,7 +27,7 @@ object ContentLength {
     }
 
   def toContentLength(value: String): ContentLength =
-    Try(value.trim.toInt).fold(
+    Try(value.trim.toLong).fold(
       _ => InvalidContentLengthValue,
       value => if (value > 0) ContentLengthValue(value) else InvalidContentLengthValue,
     )
